+++ src/cobib/config/example.py
@@ -46,6 +46,8 @@ config.commands.edit.editor = os.environ.get("EDITOR", "vim")
 
 # You can specify a custom command which will be used to `open` files associated with your entries.
 config.commands.open.command = "xdg-open" if sys.platform.lower() == "linux" else "open"
+# You can specify the names of the data fields which are to be checked for openable URLs.
+config.commands.open.fields = ["file", "url"]
 
 # You can specify a custom grep tool which will be used to search through your database and any
 # associated files. The default tool (`grep`) will not provide results for attached PDFs but other